Surface Syncing feature for the App/Sdk Views
These changes will make sure that the sandboxedSdkView
and the Ad do not have any visible ui jank between them
when being resized. We are adding SurfaceSyncGroup in
the SandboxedUiAdapterFactory's notifyResized method
which makes sure that the two surfaces are synced together
and shifting the logic of calling this method from
OnSizeChanged to OnLayout in SandboxedSdkView.

Relnote: Synchronize host and provider container size changes

Bug: 251133373
Test: manual-test
Change-Id: Ic2cd94d14cc2cc34abd3effcfa750cfccc05977e