More robust fix for createProjectZip incorrectly reporting no inputs
When Gradle runs a Zip/Copy task, Gradle appears to be caching the input directories and then not invalidating them if a subsequent task writes to those directories without declaring those writes.

Previously we worked around this issue by ensuring that the specific triggering tasks executed in a certain order.

This change implements a more robust fix that should apply to future occurrences of this problem in other projects too: it reduces the set of files that might be considered as inputs to the corresponding zip tasks, to exclude files from subprojects

Bug: 295947829
Bug: https://github.com/gradle/gradle/issues/24368

Test: # zips still not missing if run in an unlucky order:
  echo 'afterEvaluate {
  project.tasks["createProjectZip"].mustRunAfter(":compose:material3:material3-adaptive:createProjectZip")  project.tasks["compileReleaseKotlin"].mustRunAfter(":compose:material3:material3-adaptive:createProjectZip")
  }
  ' >> compose/material3/material3-adaptive/samples/build.gradle
  &&
  rm ../../out/dist/repository/androidx/compose/material3 -rf && SNAPSHOT=true ./gradlew :compose:material3:material3-adaptive:verifyOutputscreateProjectZip :compose:material3:material3-adaptive:material3-adaptive-samples:verifyOutputscreateProjectZip --no-build-cache --rerun-tasks | cat

Change-Id: I1c011d33a0016b3f012c789c2ee4028daaa2b27d