Ensure restored state is not overriden
When Compose activities are destroyed and recreated, upon restoration and new composition, compose may store state as part of rememberSavable calls. If state had already been restored at this point, the store state may override the restored state with empty bundle.

We need to ensure that restoredState is not overriden if it has already been restored.

Test: ./gradlew lifecycle:lifecycle-viewmodel-savedstate:cC
Test: manual test with sample app
Bug: 289436035
Change-Id: I629f952fc3d489cb219a3f19ff2be60d73098e01