Add maps for primitive keys and values
Bug: 285143619

Relnote: "Added common map-like classes for primitive
keys and values. For example, a key of Float and
value of Long can be stored in a FloatLongMap without
any boxing."

These classes are all generated from templates, including
the tests. Additions and bug fixes should be made to the
template and the script rerun to make sure that all
maps have the change.

Also moved primitive list and sets to be based on templates.

Test: new tests

Change-Id: Icbc204fa9dd8ef49dbfd958a2131503517a929fa