simplify-build-failure removing --assume-input-states-are-correct
Sometimes there can be mistakes in simplify-build-failure that can cause the file states psased to diff-filterer.py from necessarily passing.

For example, symlinks in the source directory aren't always handled correctly (this is fixed in a later commiti)

This change causes diff-filterer.py to re-validate that one of the given input states passes and the other one fails.

Bug: 132727976
Change-Id: I247733476c7202e48b5c480aa1508e280e24f568