Adds motion support for ThreePaneScaffold
Note: This is an initial implementation that will be used as a base for
additional changes (including API changes) and modifications that will
follow right after.

The change adds support for enter and exit transitions that are auto-
applied to the different panes at the ThreePaneScaffold.

The actual implementation is mapping the transitions that were defined
at the ThreePaneMotion to the actual pane with the current arrangement.
Then the PaneWrapper is applying an AnimatedVisibility using those
transitions.

Test: Manual
Bug: 296455429
Change-Id: I9b2e9ca41a4e4419e9661037856ba9bf0b4dbaba