Delete hyphens/line break bechnmark
The intention of the benchmark was to collect some data rather than track performance in CI

Some overall conclusion that we can make from data:
For StaticLayout construction and draw we observed (time_nanos_median stat was compared):
 - little overhead for using Heading line breaking compared to the Simple (observed up to 1% increase)
 - a slight overhead (observed up to 8% increase on some device/API combo) for using Paragraph line breaking compared to the Simple
 - additionally up to 28% overhead (median 17% for observed devices) for turning the hyphenation on (where the slower pre-API33 hyphenation algorithm used)
Please note that the exact numbers should not be referenced. This should just give a rough idea of how expensive each feature is. Also note that data is incomplete (i.e. only API 28 till 31 tested while on API 33+ the hyphenation is less costly, 128 text length and limited number of devices). Moreover, the data varies a lot from device to device.

Fixes: 247962479, 247963032
Test: N/A

Change-Id: Ibd8eed4a70a079d7b9cb1d45ea97ea2bde3d65b6