Send the correct Z-order value to the host
If the host specifies the Z-ordering prior to setting the
adapter on their SandboxedSdkView, this value should be
sent to the remote client when the session is opened.

This is tested by injecting a touch event in the location of
the SandboxedSdkView and ensuring that the touch is handled
by the correct entity in both Z-above and Z-below modes.

Relnote: Send the correct Z-order value to the host
Test: IntegrationTests
Bug: 299247180
Change-Id: Ib0ddf82ffe6942b200891e0695dcc1d961c68a2f