Merge back compat support to androidx-main
This copies over the changes from both ag/22316069 and ag/24692966 because the cherrypick tool failed. In addition to the testing below, I verified that all the modified files are identical between androidx-platform-dev and androidx-main after this change.

Bug: 299949487
Test: Run all unit tests with the following configurations:
        * U - all tests run and pass except "OlderVersions" tests, which are skipped
        * T release - skips all tests that check for ext4 or ext5, others pass
        * T with ext4 - will run as presubmit
        * T with ext5 - will run as presubmit
        * S release - skips all tests that have version checks, others pass
        * S with M-2023-10 installed via train installer - all tests run and pass except for "OlderVersions" tests and Fledge e2e
        * Same as above, but with updated webview - all tests run and pass except for "OlderVersions" tests

Note that when installing the M-2023-10 train, you need to install all of primary, telemetry, and timezone, or the sdkext version will not bump to 9.

Change-Id: Ia3e9c292d6711b32839da3920d8e60e53ae60398