Fix a crash at the DatePickers cache
The crash occured when a provided `skeleton` for a date was producing
the same `pattern` when processed by the
DateFormat.getBestDateTimePattern function (for example, when the
`skeleton` was "YY").
The fix prepend a unique character prefix for the skeleton and the
pattern keys used with the cache.

Bug: 301289644
Test: CalendarModelTest
Relnote: "Fix a crash when the DatePicker is initialized with a
DatePickerFormatter that has certain date skeletons that stay the same
when converted to date patterns (such as YY)."

Change-Id: I01f2902849d4dd5c2426f4d3c13f52463692e3f4