Fix unexpected exit events on mouse click
A mouse click sometimes emits an Exit event. This was originally handled in aosp/1875579, but mouse events were omitted since they usually were paired with a button press. My mouse emits an exit motion event without a button press before the press event, so this reorganizes the logic to check mouse events unless a button press is included in the motion event.

Test: Manually verify Logitech M557 mouse clicking on a TextField does not change pointer icon
Test: ./gradlew :com:ui:ui:tDUT
Test: ./gradlew :com:ui:ui:cAT
Bug: b/283122266
Change-Id: I6df393c57a81c4851779cf116315b19ea4b0d581