Refactoring @RequiresApi annotation to be more KMP-friendly.
The @RequiresApi annotation has been replaced with @RequiresAndroidApi
annotation in commonMain and a typealias has been added to jvmMain to
continue supporting the usage of the original api.

This change allows the Api version specific annotation to stay within
the Android side of things, with the option to later on add an iOS
specific annotation in the future.

Bug: b/302156663
Test: Existing Tests
Change-Id: I3195908be7be01c3e6b85ef82cad1cd4dc9ef17e