Avoid over-sanitizing filenames breaking BP plugin file filtering
Fixes:303034735
Bug:296453339
Test: OutputTest
Relnote: "Fix Benchmark file output to no longer break BaselineProfile
Plugin file copying. Files were generated and copied off device, but
had been renamed such that the gradle plugin wouldn't see them."

The previous change to fix b/296453339 sanitized filenames to avoid
certain characters (e.g. the equals sign which caused issues in CI /
instrumentation result reporting), but unintentionally broke a load
bearing assumption between BPGP and the Benchmark library.

This wasn't caught in testing because continuous test are on either
the host side or device side, and don't cover the exchange of files in
between.

To workaround this coverage gap for now, added a test to OutputTest to
cover the BPGP filtering - duplicating the filtering that BPGP does to
look for baseline and startup profiles.

Change-Id: I8dbccc5a11b4c391169cff6dfbd2a0001fd41001