Moved PlatformTextInputMethodRequest into the same module as the rest of PTI.
It used to be in ui-text, everything else is in ui. I think that was
because there was other code in there from a few refactors ago.

The advantage of having them all in the same module is that we can
introduce experimental/internal stuff in the future if we need to.
It's also just overly complicated and inconsistent to split this API
across multiple modules unnecessarily.

Relnote: n/a
Test: n/a
Change-Id: Ib0e800ecfc4ec8db6b5179c764717781486118da