Fix the logic of acting handle while long press dragging on BTF2
This change mainly addresses the problem with the logic behind deciding the acting handle while dragging after long press gesture is being applied on BTF2. It's called acting handle because there are actually no handles being dragged or interacted with in this scenario. However, we would have to assume one of the handles are being carried to address where the magnifier would go, or how the selection would be adjusted.

New logic deciding the acting handle is very simple. At each selection change while the drag events are processed, we check whether the start or the end of selection has changed. If they have both changed, we follow the middle of the selection. Finally if the selection hasn't changed at all, we do not update the acting handle unnecessarily. Additionally, reverse selection through long press drag gesture is now blocked due to a visual glitch that happens because selection never collapses on the initially selected word.

Furthermore, this change adds extensive comments to explain each peculiar looking piece of code. It's really important to explain the reasoning behind certain behavior effecting decisions for future maintainability in a class like TextFieldSelectionState.

Test: N/A
Change-Id: I41dcefa8ced25131c338aa00e4a14ec56a279962