Add support for Multi-Resolution reprocessing
Change CameraGraph.Config to accept a list of InputStreams instead of one InputStream.

In CaptureSessionFactory, pass in the list of InputStreams to createCaptureSession when supported. Update calls to CameraGraph.Config.input to treat parameter as a list instead of a single object.

Update createCaptureSession to have ability to set multi resolution InputConfiguration.

Add validation checks on API level and number of InputStreams in CameraGraphImpl.

Bug: 304344851

Test: ./gradlew --info --strict \
 :camera:camera-camera2-pipe:testDebugUnitTest \
 :camera:camera-camera2-pipe:compileDebugUnitTestKotlin \
 :camera:camera-camera2-pipe-testing:testDebugUnitTest \
 :camera:camera-camera2-pipe-testing:compileDebugUnitTestKotlin \
 :camera:camera-camera2-pipe-integration:testDebugUnitTest \
 :camera:camera-camera2-pipe-integration:compileDebugUnitTestKotlin \
 :camera:integration-tests:camera-testapp-camera2-pipe:testDebugUnitTest \
 :camera:integration-tests:camera-testapp-camera2-pipe:compileDebugUnitTestKotlin

Change-Id: I62ecf03014fb17b614c25517a184cfc3130e0641