Rename FileDescriptorIoException to DataInputException for Alpha-07 release
Renaming user-facing error to be more generic instead so that it does
not highlight implementation details. The server side exception
remains the same (must do due to stableAidl). The new name suggests
that there is a problem with the way the data is being input, which
is generic enough to also cover our future plans to support inputStreams
as an input instead of just file descriptors.

Test: ./gradlew :javascriptengine:javascriptengine:connectedAndroidTest --info
Bug: 286615957
Relnote: Rename FileDescriptorIoException to DataInputException for Alpha-07 release

Change-Id: Iba4ebd392f3e4ca7c0a4831a1a76fa5635a78596