Add ability to customize LazyLayout prefetching strategy
Our goal is to enable testing of different prefetching strategies, including what to prefetch and when that work is scheduled. This diff refactors the existing logic behind a strategy pattern, maintaining the current choices of how much/when as the default strategy.

The existing logic is split between LazyListState (+ LazyGridState, etc), which decides *what* to prefetch, and LazyLayoutPrefetcher, which decides *when* to execute it.

A summary of changes:
- LazyList accepts an optional LazyListPrefetchStrategy falling back to DefaultLazyListPrefetchStrategy which uses the existing logic.
- LazyListPrefetchStrategy contains LazyList specific callbacks that a strategy can use to decide how to manage prefetching (currently onScroll and onVisibleItemsChanged)
- LazyListPrefetchStrategy extends the LazyLayoutPrefetchStrategy interface, which allows LazyLayout to connect a PrefetchContext to the strategy for it to execute prefetches. This is analogous to how LazyLayout connects the LazyLayoutPrefetcher to a LazyLayoutPrefetchState before this diff.
- PrefetchContext gives access to the precompose API from SubcomposeLayout instead of exposing how to scheduling a prefetch - this was done to allow a strategy to determine its own scheduling logic
- LazyLayoutPrefetcher mostly stays as is, except it's given a PrefetchContext instead of setting/unsetting itself in onRemembered/onForgotten and is directly used from the default prefetch strategy.

I didn't try to implement correct logic for anything besides LazyList and didn't clean up LazyLayoutPrefetchState, just to get feedback on this approach first.

I also uploaded https://android-review.googlesource.com/c/platform/frameworks/support/+/2817620 as an example of custom logic using the API to change the amount to prefetch and when to execute prefetch.

Test: * Rotate sample app with custom PrefetchStrategy. LazyList still uses chosen PrefetchStrategy after rotation.
* Run TrivialListScrollBenchmark without custom strategy, see no changes
* Run TrivialListScrollBenchmark with custom strategy/executor, see potential perf wins when rows can take over a frame to execute.
* Unit tests in this patch

Relnote: Adds an experimental API for configuring prefetch behavior of LazyLists.

Change-Id: I022a469c6d89d8742511cf10715afeced732d27e