Remove unreliable draw timing assertion from PhaseOrderingTest.
Compose doesn't actually control when draw happens, and this test that
asserted strongly on when it happens in relation to other compose phases
was flaky on API 21. Since we don't control it, it doesn't make sense to
assert on it, so I just removed the draw ordering assertion.

Test: PhaseOrderingTest
Bug: b/298694432
Relnote: n/a
Change-Id: I493ddc046e58ac2debaeb7f196320759e1709a8e