Fix NavDestination equality when using NavAction
NavAction didn't override equals() or hashCode()
so it was using reference equality. That meant that any
NavDestination that included an action would never be
equals() to one another because:

1) NavAction was only using reference equality
2) SparseArrayCompat's containsValue uses reference equality

So we needed to change both NavAction (which now
has equals(), hashCode(), and toString()), NavOptions
(which now has a toString() which is used by NavAction),
and NavDestination (to actually NavActions by equals()).

Relnote: "Fixed an issue where your back stack would be
popped if you called `setGraph` more than once with the
exact same graph only if there was a destination in your
graph that contained an action linking two destinations."
Test: updated tests pass

Change-Id: Ieaed7edf02f70696cd0a3558cfe8d7960d1cddd4