Fix handling deep links only once on NEW_TASK+CLEAR_TASK
handleDeepLink() needs to set the deepLinkHandled flag
if it handles a deep link, but it was only doing this if
the Intent didn't have NEW_TASK in it (i.e., you were on
another app's task stack). We need to set the flag in both
cases.

Relnote: "Fixed an issue where NavController could
handle the same deep link more than once if the back
stack was entirely empty before a configuration change
or call to `setGraph` only when the incoming Intent
had the `FLAG_ACTIVITY_NEW_TASK` flag set."
Test: updated tests pass

Change-Id: I73c7f78a55eb283ac4b2fa7b10f6ea05a8fd8db7