Final tying bits for clang compilation infra
This CL introduces APIs to setup native compilation from build.gradle
files and also include their outputs in klibs, jars and aars.

New tasks:
* CombineObjectFilesTask: This is a simple task that receives a list of
object file inputs and organizes them in a well defined directory
structure.
The output directory can either be included as jniLibs or
put into Jar resources.
* CreateDefFileWithLibraryPathTask: This is a special task that will
produce a CInterop def file which also includes the path to a given
archive library.
We use this task to configure cinterop for kotlin native targets such
that the native library will be included in the final klib file.

There are also 2 helper classes using these tasks:
* MultiTargetNativeCompilation.configureCinterop: Configures CInterop
for a kotlin native target based on the outputs of a
MultiTargetNativeCompilation.
* NativeLibraryBundler: Configures Android Variants or Jars to include
the output of MultiTargetNativeCompilation.

All public API for native compilation is exposed via the
AndroidXMultiplatformExtension.
* createNativeCompilation: Creates a new setup for native compilation
that can be bundled with final targets
* createCinterop: Sets up the library to embed the native library and
generate its cinterop stubs
* addNativeLibrariesToJniLibs: Adds the AndroidTargets of the native
compilation to the jni libs.
* addNativeLibrariesToResources: Adds the desktop-os targets of the
native compilation to the jni libs.

Test: CombineObjectFilesTaskTest, CreateDefFileWithLibraryPathTaskTest

Change-Id: I16f62721dc2562d3d517bc356af29aa3b3003311