Update clip demo to include unclipped ghost text
The text used in the demo is shortened so that it doesn't take as much room when displaying the whole text. The text then is underlayed below the clipped text with decreased alpha and no clipping. This allows us to view where the clipped text would be put if it weren't clipped.

Test: N/A
Change-Id: Ibaaa279ea072c05728a80ebf1a29b64afd6306b4