Convert ActivityResultLauncher to Kotlin
Rewrite ActivityResultLauncher in Kotlin.

The getContract() method is now an abtract Kotlin
property.

This also moves the `ActivityResultLauncher` Kotlin
extensions from `activity-ktx` back to `activity` since
we can't have two Kotlin files with the same name.

Relnote: "`ActivityResultLauncher` has been rewritten in
Kotlin. As part of that conversion, the `getContract` method
is now an abstract Kotlin property. This is a binary compatible
change, but source breaking if your implementation of
`ActivityResultLauncher` is written in Kotlin."
Test: ./gradlew activity:activity:checkApi
BUG: 257291701

Change-Id: Id4615d19b383b2438be6a8f68da7584394d13515