Prevent catching the switch while animating on quick clicks
When clicking on the switch twice fast enough, it causes "catching" the switch while it is animating to one side. This is a result of passing `startDragImmediately` as true to Draggable when switch is animating. To prevent this from happening, `startDragImmediately` is exposed for anchoredDraggable and the switch now passes a false boolean to it. This will prevent catching the animating switch by only pressing on it.

Fixes: 215768437
Test: Manual test
Change-Id: Ic3a4893fedfa60261d7eca7c576d9e168a1ee788