Address API council feedbacks for ScreenFlashUiControl naming and timeout constant
As per the feedbacks and the recommendations, the following changes have been done:
  1. Simplify ScreenFlashUiControl and its member method names to ScreenFlash, ScreenFlash#apply(ScreenFlashUiCompleter) and ScreenFlash#clear()
  2. Instead of exposing SCREEN_FLASH_UI_APPLY_TIMEOUT_SECONDS directly, it is exposed via a getter

This CL only renames the APIs and updates the documentations without changing the api folder files. The api files will be changed in another CL after this to make it easier to review the API surface changes.

Bug: 306690224
Test: ./gradlew bOS
Change-Id: I9e7c4bddb044b4ce6241ae71f6ecfb739bc6f6c8