Ensure adding callback doesn't stop current dispatch
Currently, if you add an OnBackPressedCallback to the
OnBackPressedDispatcher after your have already dispatched onBackStarted
to a previous callback, the new callback will receive the
onBackProgressed and onBackPressed calls and the previous callback will
never complete.

We need to make sure that when a callback gets onBackStarted, we always
complete that callback regardless of if a new one is added.

RelNote: "`OnBackPressedDispatcher` now continues to dispatch to the
correct `OnBackPressedCallback` even after a new `OnBackPressedCallback`
is added while the back gesture is being handled."
Test: added new tests
Bug: 303481184

Change-Id: Id0ff6b50a2a7f721fc13bab5ca63f620f1e3e25e