Fix flaky VideoCaptureTest
The issue is somehow the cleanup code posted on the main thread in SurfaceProcessorNode#release() is not executed. The unit tests do wait for the main thread to become idle so I am not sure why it's not happening.

The fix is run the cleanup code directly if it's already on the main thread. We should do this anyway to reduce thread hopping. After the fix, running through all unit tests no longer produces the "garbage collected" error.

Bug: 309409701
Test: manual test and ./gradlew bOS
Change-Id: I35f12741838c5010f931527fdb3a485f499d5edb