Perform exact computations in CubicBezierEasing
The existing implementation of CubicBezierEasing was performing a
binary search, evaluating the Bézier at multiple points to find a
position on the curve that would yield an X position close enough
to the easing function's input fraction

The new implementation performs an exact computation (within the
limits of precision afforded by double precision computations on
a CPU) instead. First, we find the real roots that solve the
cubic polynomial that defines the curve. Then using one of the
roots (we just take the first one), we can simply evaluate the
cubic function and return the Y value.

To find the root at any given X coordinate along the curve, the
curve is first shifted towards the origin by the input fraction.
Solving for Y=0 in the shifted curve gives us the solution we are
looking for.

The code we use here will later serve as the basis for a more
complete solution that will allow us to evaluate a PathEasing
curve at any X without any approximation. In addition to find the
roots for and evaluating cubic Bézier curves, we'll need to do it
for quadratic Bézier curves and lines. Thankfully these are
trivial compared to cubics.

It is important to note that the evaluateCubic() function
provided here takes a shortcut by assuming that the start and end
points are respectively 0,0 and 1,1 since this is mandated by
CubicBezierEasing. To solve for arbitrary paths, we'll need a
more general evaluateCubic() function that will be added later.

This change also adds a useful test to CubicBeziserEasing. The
only tests were only testing for equals() and the start/end
points, without testing for the computations performed in
transform() at arbitrary fractions. To close this gap, we
introduce an instrumented test that uses the new implementation
to manually rasterize cubic curves in a Bitmap: for each X
coordinate in the Bitmap, compute the fraction along the X axis
and solve the curve using transform(). Using the resulting Y
coordinate we can draw a line from the previous X/Y pair and
draw the curve.
By doing this, we can perform a pixel-to-pixel comparison using
a Bitmap containing a Path rendered via Canvas. We must allow
for a certain amount of error since our computation will be
different from the ones performed by Canvas (mostly because
Canvas itself uses a subdivision approximation to render a
Path). This error is defined as a total number of pixels that
are allowed to be different between the two bitmaps.

Test: EasingTest (instrumented and host)
RelNote: CubicBezierEasing now uses exact solutions
Change-Id: I2da4dc1e64c28124e4eb7fd8e05950089f191e29