Relax frame structure requirements API < 31
Bug: 307845882
Test: FrameTimingQueryTest
Relnote: "Relaxed strictness of checks in FrameMetrics, and added more
detail to error messages"

It's possible these are caused by devices that don't conform to
standard frame tracing, possibly due to cherry-picks of 31 tracing to
older devices. As this should be a benign error, we avoid throwing in
this scenario.

Additionally, convert require -> check since these aren't strictly
argument failures, and add additonal data to the remaining checks,
which request a perfetto trace.

Change-Id: Iadededa9b5721f1e68534a0457c3dfc2f8c91b0f