Use exact computations in PathEasing
Prior to this change PathEasing was approximating the easing
path by flattening it as lines. However, this caused multiple
problems:

- The flattening process requires an error metric that should
in theory depend on the total duration of the animation.
Since the duration is not available to the easing curve, the
implementation had to pick a constant approximation that was
too fine-grained for more animations (thus requiring to
subidivide the path more than needed) and too coarse for
long animations (creating large linear approximations of the
actual easing curve).

- The flattening process was unnecessarily based on the path
length. Instead of chopping up the path along the X axis,
PathEasing was first looking for the total length of the path
to then find points at regular intervals along the path. This
could cause also unncessary subdivision in areas with steep
slopes, but more importantly that code was harmed by an API
limitation: PathMeasure.length does not allow to select the
error metric.
Since length is itself based on a subdivision
process, it should in theory use the same error as the
approximation as lines. Unfortunately the default error
metric for PathMeasure.length is 1/2, which makes absolute
sense when measuring paths to draw them on screen (half a
pixel gives a reasonably good approximation to render the
path as a series of lines). But in the context of an easing
curve, this error metric is *terrible*. The curve is
always defined in the 0..1 domain, which means the error
metric of 1/2 leads the path measurement code to always
assume the curve is a straight line from 0,0 to 1,1, thus
always returning a fixed length of sqrt(2). This sole
issue can cause the length to be off by ~30% with a
single "fast in slow out" curve for instance. This dif-
ference means that the lines approximations themselves
can be misealding as the entire path might not be sampled
at all.

These two issues combined meant that the existing
implementation would always use 500 linear approximations
no matter how long the animation nor how complex the path.

The new implementation takes an entirely different approach
and instead inspects the content of the path to find the
curve that should be sampled for any time fraction.

Before describing the evaluation of the curves, lets
describe how the path is inspected then queried:

- The implementation uses the newly introduced
Path.iterator() API to iterate over every segment of the
path. Some of those segments are ignored (Done/Move) and
some are forbidden (Close).

- For each segment, we compute the bounds of the segment's
curve. To achieve this, we must look at the derivative
and potentially second derivative of the curve, find the
roots of the derivatives and evaluate the curve at the
positions defined by those roots. This gives us the
extremas for the curves, which can be min/maxed to find
the bounds (note: we only care about the bounds along
the X axis in our case).

- The bounds of the segment give us an interval along
the X axis that we use to build an interval tree. The
interval tree has two interesting properties for our
use case:
  1. We can quickly and easily query which curve(s)
     contain the input fraction from inside transform()
  2. We can easily handle multiple curves overlapping
     on the X axis

- After building the interval tree, we can run a couple
of simple validity checks to see whether the path
starts at x=0 (we actually test if one of the curves
contains x=0), and ends at x=1 (same).

We the interval tree, the implementation of the
transform() function is fairly straightforward:

- Given an input fraction, we query the tree for any
curve that contains that fraction in the X axis.

- Once we have a curve, we must solve it for
x=fraction. We do so by offsetting the curve on the
X axis by -fraction and solving for y=0. This gives
us a position t along the curve for a given x.

- With the t value we just computed, we evaluate the
curve at that position to find the value on the Y
axis and return that as the result of our transform.

Most of the complexity in the code lies with dealing
with 3 types of curves:
  1. Lines
  2. Quadratic curves
  3. Cubic curves

With this new implementation, the precision is not limited
by an arbitrary constant but by the precision of our
computations (we use doubles to minimize the precision
loss). The result is also completely independent of the
total duration of the animation since we never approxi-
mate the curves.

RelNote: PathEasing now uses exact, duration independent computations
Test: PathEasingTest, integration demos
Change-Id: I1b327f124225e7896db1182885ee2ea24c85af56