Create new LifecycleCamera and CameraUseCaseAdapter instance for new CameraConfig.
Currently, LifecycleCamera which contains a CameraUseCaseAdapter
is one instance per (LifecycleOwner, CameraInternal) combination.
This works fine until Extensions introduce restricted version of
CameraInfo and CameraControl. The restriction that Extensions
introduce will be reverted right after apps binds with non-extensions
CameraSelector due to the there is just one CameraUseCaseAdapter
for the same CameraInternal when LifecycleOwner is the same.

E.g., if apps bind with the Extensions-enabled CameraSelector,
the CameraControl / CameraInfo from the returned Camera might
disable zoom or focusMetering depending on OEM capabilities.
If later apps bind again without enabling the extensions but with
the same camera id,  then it will get the regular CameraControl
/CameraInfo and futhermore, for the previous CameraControl/CameraInfo's
of the Extensions-enabled camera, restriction will be removed.

To solve this issue, we need to create new instance of
LifecycleCamera and CameraUseCaseAdapter.  The CL is to implement
the new behavior which includes

1. Add CameraConfig.getCompabilityId to the key of
   LifecycleCameraRepository and create new instance of
   LifecycleCamera and CameraUsecaseAdapter when binding with new
   CameraConfig.

2. make CameraConfig
   immutable to the CameraUseCaseAdapter.  This is no longer
   necessary for CameraUseCaseAdapter to dynamically change
   CameraConfig because different CameraConfig will creating
   corresponding CameraUseCaseAdapter instances.

3. ensure CameraConfig is configured correctly in the
   CameraInternal when switching between LifecycleCamera.

Bug: 312662243
Test: LifecycleCameraRespositoryTest, CameraUseCaseAdapterTest
Change-Id: I3ac2e2e31f4ba7b35dae73326af5bdf83da608f2