Fix for ignored lazy benchmarks
After the recent optimization where we stopped doing the synchronous remeasure some of the benchmarks started to fail as the remeasure wasn't happening with the previous logic. View.measure() was skipping the work as no constraints were changed. Now we will use our  mechanism of performing remeasure/relayout we use in tests as well.

Test: run LazyListScrollingBenchmark
Fixes: 311256618
Change-Id: Ifdf594f3569fa121fd6f5e525e6b51f54f759af2