Adds RippleConfiguration APIs to material3
RippleConfiguration and LocalRippleConfiguration allow for per-component / sub-tree customization of ripples using fixed values. For example, to change the color of a component you don't control, or to disable a ripple for a component. In most cases the default values should be used: these APIs are an escape hatch for customization of individual components / limited sub-trees. For wider changes and custom design systems, you should instead build your own ripple using createRippleModifierNode.

Bug: b/298048146
Test: RippleTest
Relnote: "Adds RippleConfiguration and LocalRippleConfiguration to allow for per-component / sub-tree customization of ripples using fixed values. For example, to change the color of a component you don't control, or to disable a ripple for a component. In most cases the default values should be used: these APIs are an escape hatch for customization of individual components / limited sub-trees. For wider changes and custom design systems, you should instead build your own ripple using createRippleModifierNode. For more information, see developer.android.com."
Change-Id: I7b5d62fd50ee78bb3559f83886aa1e7d9f964fb1