Add SessionManager.runWithLock
Refactor SessionManager so that all operations are protected by a
lock, and multiple operations can be run while holding the lock. This
prevents a race between different SessionManager clients, who want to
start a session if it is not currently running, from both trying to
start a session at the same time.

For example, given 2 clients (GlanceRemoteViewsService and
GlanceAppWidget), that want to start a session, the following behavior
is currently possible:

Client 1			Client 2
isSessionRunning() = false
       				isSessionRunning() = false
startSession(s1)
       				startSession(s2)
s1.doEvent() // throws because s1 was closed by startSession(s2)

After this change, clients are forced to do operations within
runWithLock, which prevents the interleaving above:

Client 1			Client 2
mutex.lock() // successful
				mutex.lock()	// suspends
isSessionRunning() = false
startSession(s1)
s1.doEvent()
mutex.unlock()
				isSessionRunning() = true
				getSession() = s1
				mutex.unlock()

Tests are also refactored to use the new API.

The waitForReady event is refactored to return a Job that can be
`join`ed to wait for the session to be ready. This allows clients to
wait for session ready outside of the runWithLock critical section.

Bug: 311428322
Bug: 308760288
Test: SessionManagerImplTest.runWithLockIsMutuallyExclusive
Test: AppWidgetSessionTest.waitForReadyResumesWhen{EventIsReceived,SessionIsClosed}
Change-Id: I87496389a1ef3f159dac787cdc82ca1f38c6f47e