Optimize onLayoutChange in accessibility delegate
`isEnabled` checks `DisableContentCapture` feature flag, recording state reads for each placed node. This is wasteful, as we are not interested in a fine-grained observation in this case (flag is usually flipped at the start of the activity/app)

Also replaces `boundsUpdateChannel` to be `Channel(1)` which is significantly faster than conflated channel as it doesn't have to replace the value.

Test: Existing tests
Change-Id: Ic92ef2b8cdd6474424651c5374c3d84495d5c51a