Do not refer to restricted class in docs.
This commit updates a few references to the restricted classes
ComplicationDisplayPolicy and ComplicationPersistencePolicy, as these
are used in public documentation but they are not themselves public.
These actually do not need to be public, and we can instead refer to
the classes that define the constants that can be used to express
these.

Bug: 314072967
Test: Manual build
Change-Id: I4a6137e81c9cfd8af5ce6cef40c36304d0dff937