Fail quickly when KSP does not completely run in testing infra.
If KSP crashes before the synthetic processor is initialized, the test will crash due to an 'uninitialized property' while the real exception is lost in the diagnostics messages. This change checks that KSP runs as expected and if not throw a more useful exception containing the compiler messages founds in the collector.

Bug: 314038891
Test: n/a
Change-Id: I35ccd4ec762b5eca0478545524a650634a39a239