Exclude font padding by default on Compact and Title chip
This changes the default behaviour of these two chips to
exclude font padding always and to remove existing setter
for it.

Bug: 314728587
Test: Existing passing because there's no visual diff in these chips
Relnote: "Following Compose and Wear Compose, we are making a change to turn off the font padding by default in all ProtoLayout and ProtoLayout Material components. This will address some cases of text clipping with large font sizes and may also impact screen layouts, so that screenshot tests need updated. As this fixes problems with misaligned label, we are removing the setters for it."
Change-Id: I329598c92ed34d0ee80ce2db7d2d7f021d2c6883