Add SQLiteException to SQLite KMP
This change adds an exception class in common code to be used by drivers to signal errors. This change also adds a few more conformance tests to verify exception and perform checks on binding and getting column APIs. Specifically check there is a row to get data out of and check range of binding and getting columns indexes.

Additionally:
 * Align usage of UTF-16 C APIs in native driver as Kotlin native String are UTF-16 encoded.
 * Opt-in to ExperimentalForeignApi in code (in addition to compiler flag) to improve IDE developer experience.
 * Add a script for running conformance tests in the drivers.

Bug: 304297717
Bug: 307917224
Bug: 311507338
Test: ./sqliteMultiplatform/runConformanceTest.sh
Change-Id: I3921ce2c036c5bd934fc9e3dbc1acad9a17f1f95