Fix VideoCapture incorrect cropping due to input stream rotated
Previously, the camera stream flowed to VideoCapture has the same orientation as the camera sensor. However, after aosp/2847734, StreamSharing now rotates the stream before feeding it to VideoCapture. This means the suggested resolution might have a larger height than width. When using the MediaCodecInfo API, specifically VideoCapabilities.isSizeSupported(width, height), the suggested resolution is used as the parameter. By experimentation, it was found that MediaCodecInfo treats the longer side as the width and the shorter side as the height. Therefore, reversing the order of width and height might lead to isSizeSupported() returning false. For example, isSizeSupported(3840, 2160) returns true but isSizeSupported(2160, 3840) returns false. When adjusting the crop rect based on this incorrect size constraint could result in inaccurate cropping.

For now there is no API to know if the codec support swapping width and height. By experimentation, almost all default codecs support swapping width and height for encoding.

This CL
* Add (internal) API VideoEncoderInfo.canSwapWidthHeight() to query if a codec allow swapping width and height.
* Add (internal) API VideoEncoderInfo.isSizeSupportedAllowSwapping(width, height) to make it esay to check size when canSwapWidthHeight() is true.
* Add SwappedVideoEncoderInfo to make it easy to adjust the crop rect in VideoCapture.

Also refactored in this CL:
- Way to add extra supported size of VideoEncoderInfoWrapper.
- Apply VideoEncoderInfoWrapper in the default EncoderInfo finder.

Bug: 314694668
Bug: 299420960
Test: ./gradlew camera:camera-video:connectedAndroidTest
Change-Id: I12b4c05662292aa104df059d106d9651dd510aab