Moves hover and focus nodes inside Clickable
This is a halfway step for flattening the delegation tree inside clickable and unblocks some near term optimizations. In the future delegation should be avoided here and the logic should be inlined to reduce the size of the delegated node tree.

This CL also fixes an issue where focus properties were leaking down the tree when clickable was disabled.

Fixes: b/269319898
Test: ClickableTest
Test: CombinedClickableTest
Change-Id: I5bf797a08faff79f690e2a8d3c57a4e4a67f9dce