Improve error logging when metric failures occur
Bug: 315170517
Test: SwipeBenchmark (currently failing)
Test: PerfettoTraceProcessorTest
Relnote: "Improve FrameTimingMetric error when frames aren't produced,
and always output link to trace when failing metric parsing to assist
in diagnosing problem."

Further improvement could be made by getting rid of FileLinkingRule
and folding the functionality into
PerfettoTraceProcessor#loadTrace. This would ensure traces are logged
on success as well, but that requires changing the API and all of the
common callers.

Change-Id: I956b90505fbd72a80dc00ad2214420f4d3957659