Don't throw from isCurrentlyConstrained in case of ConstraintTrackingWorker
ConstraintTrackingWorker could be run on API 28 due to OS upgrades.
In this case we are currently throw from isCurrentlyConstrained.
This CL stops throwing if the WorkSpec doesn't have requiredNetworkRequest
and simply returns that workspec isn't constrained. This will resolve the
issue because ConstraintTrackingWorker are only enqueued on API 23..25,
and as results these workers won't have requiredNetworkRequest.

bug: 314322766
Test: NetworkRequestConstraintControllerTest
Change-Id: Iec7cc02871fba4ed50fefb851f7d15513b3a2b57