Create a databaseBuilder function that takes in an factory to avoid reflection.
The logic behind the factory is to create access to the generated database implementation for the user on the Native side. This functionality has been enabled on the Android side as well, however the Android RoomDatabase builder will require a Context (which the Native implementation does not). There is a databaseBuilder() on the Native side that takes in a factory, and there is a version of the same function on the Android side that takes in a factory as well as a Context.

Since this is a Kotlin-only functionality, the `instantiateImpl()` function is only generated when Kotlin codegen is on.

This CL also introduces an initial BuilderTest in the nativeTest, which will have additional tests added once the DatabaseConfiguration has been fully moved to commonMain.

Bug: 316978491
Test: ./gradlew :room:integration-tests:room-testapp-kotlin:kspWithKspGenKotlinDebugAndroidTestKotlin
Relnote: Create a databaseBuilder function that takes in an factory to avoid reflection.
Change-Id: I5f5be211f5752733f113ef0fc3f5772866ca84c8