Remove lowVelocitySpec and update SnapLayoutInfoProvider approach to have a default value.
In this change we're removing the need for a lowVelocitySpec in SnapFlingBehavior. The use case for low velocity flings will be taken care of by the snapAnimationSpec. This change also updated the approach method in SnapLayoutInfoProvider to have a default value, therefore no longer being required. If an approach value isn't provided, SnapFlingBehavior will try to decay by default.

Test: Previous tests should pass.
Relnote: In this change we're removing the need for a lowVelocitySpec in SnapFlingBehavior. The use case for low velocity flings will be taken care of by the snapAnimationSpec. This change also updated the approach method to have a default value, therefore no longer being required. If an approach value isn't provided, SnapFlingBehavior will try to decay by default.

Change-Id: I3c0adbec079b83a3060548acf1185b90a13132bb